- Move useAuth to shared folder. - Updated AuthProvider for website and added for mobile.
