+++ Mobile/app/_layout.tsx
@@ -3,7 +3,8 @@ import { SafeAreaProvider } from "react-native-safe-area-context";
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
 import { KeyboardProvider } from "react-native-keyboard-controller";
 
-import { ThemeProvider } from "@/providers/ThemeProvider";
+import { ThemeProvider } from "@/providers/shared/ThemeProvider";
+import { AuthProvider } from "@/providers/app/AuthProvider";
 
 export default function RootLayout() {
   const queryClient = new QueryClient({
@@ -19,10 +20,12 @@ export default function RootLayout() {
       <ThemeProvider>
         <SafeAreaProvider>
           <KeyboardProvider>
-            <Stack>
-              <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-              <Stack.Screen name="(auth)" options={{ headerShown: false }} />
-            </Stack>
+            <AuthProvider>
+              <Stack>
+                <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+                <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+              </Stack>
+            </AuthProvider>
           </KeyboardProvider>
         </SafeAreaProvider>
       </ThemeProvider>
