+++ Mobile/src/providers/app/AuthProvider.tsx
@@ -1,12 +1,11 @@
 import { axiosInstance } from "@/api/axiosInstance";
-import { Loader } from "@/components/app/Loader";
 import { AuthContext, useAuthProvider } from "@hobbyist/hooks";
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const { value, isPending } = useAuthProvider(axiosInstance);
 
   if (isPending) {
-    return <Loader />;
+    return <></>;
   }
 
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
