+++ Shared/form-schemas/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    /* Environment */
+    /*Environment */
     "target": "ES2020",
 
     /* Modules */
@@ -17,8 +17,13 @@
     /* Interop Constraints */
     "composite": true,
 
-    /* Type Checking */
+    /* Linting */
     "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "erasableSyntaxOnly": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedSideEffectImports": true,
 
     /* Completeness */
     "skipLibCheck": true
