+++ Shared/hooks/tsconfig.json
@@ -17,8 +17,13 @@
     /* Interop Constraints */
     "composite": true,
 
-    /* Type Checking */
+    /* Linting */
     "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "erasableSyntaxOnly": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedSideEffectImports": true,
 
     /* Completeness */
     "skipLibCheck": true
